Moving vim state into doceditor, centralizing vim to doceditor
