+++ novelwriter/guimain.py
@@ -44,14 +44,13 @@ from novelwriter.dialogs.about import GuiAbout
 from novelwriter.dialogs.preferences import GuiPreferences
 from novelwriter.dialogs.projectsettings import GuiProjectSettings
 from novelwriter.dialogs.wordlist import GuiWordList
-from novelwriter.enum import nwDocAction, nwDocInsert, nwDocMode, nwFocus, nwItemType, nwView
+from novelwriter.enum import nwDocAction, nwDocInsert, nwDocMode, nwFocus, nwItemType, nwView, nwVimMode
 from novelwriter.extensions.progressbars import NProgressSimple
 from novelwriter.gui.doceditor import GuiDocEditor
 from novelwriter.gui.docviewer import GuiDocViewer
 from novelwriter.gui.docviewerpanel import GuiDocViewerPanel
 from novelwriter.gui.itemdetails import GuiItemDetails
 from novelwriter.gui.mainmenu import GuiMainMenu
-from novelwriter.gui.vimstate import VimState, Mode
 from novelwriter.gui.noveltree import GuiNovelView
 from novelwriter.gui.outline import GuiOutlineView
 from novelwriter.gui.projtree import GuiProjectView
@@ -182,11 +181,6 @@ class GuiMain(QMainWindow):
         # Editor / Viewer Default State
         self.splitView.setVisible(False)
         self.docEditor.closeSearch()
-
-        # Vim state for vim mode
-        self.vim = VimState()
-        self.vim.enabled = CONFIG.vim_mode
-        self.docEditor.setVimModeState(self.vim)
 
         # Progress Bar
         self.mainProgress = NProgressSimple(self)
@@ -1332,8 +1326,7 @@ class GuiMain(QMainWindow):
     @pyqtSlot()
     def _keyPressEscape(self) -> None:
         """Process an escape keypress in the main window."""
-        if self.vim.enabled:
-            self.vim.set_mode(Mode.NORMAL)
+        self.docEditor.setVimMode(nwVimMode.NORMAL)
 
         if self.docEditor.searchVisible():
             self.docEditor.closeSearch()
